feat(CarController): 인터페이스 정의 후 CarControllerImpl로 수정
