+++ src/main/java/racingcar/Application.java
@@ -1,9 +1,6 @@
 package racingcar;
 
 import racingcar.domain.AppConfig;
-import racingcar.domain.car.dao.CarRepository;
-import racingcar.domain.car.controller.CarController;
-import racingcar.domain.car.service.CarService;
 
 public class Application {
     public static void main(String[] args) {
